Further porting of PixelsService.
This port is the upcasting of sizeX * sizeY in the context of
calculating whether or not a pyramid is required for the Pixels set in
question.  Without this change we have an integer overflow and images
that require a pyramid are not assigned one and the client is confused,
asking for whole plane rendering where the total number of pixels are on
the gigabyte scale.

+++ components/romio/src/ome/io/nio/PixelsService.java
@@ -455,8 +455,8 @@ public class PixelsService extends AbstractFileSystemService
      * @return
      */
     public boolean requiresPixelsPyramid(Pixels pixels) {
-        final int sizeX = pixels.getSizeX();
-        final int sizeY = pixels.getSizeY();
+        final long sizeX = pixels.getSizeX();
+        final long sizeY = pixels.getSizeY();
         final boolean requirePyramid = (sizeX * sizeY) > (sizes.getMaxPlaneWidth()*sizes.getMaxPlaneHeight());
         return requirePyramid;
     }
